Add again slow method
It produces results that are correct, is optimized enough to avoid stack overflows on the test cases, but does not terminated within the time limit on all test cases.

Its main benefit is to allow investigate why the eager method does not work by making it possible to compare the two.